stripping it down more -
removed timeouts after carefully moving them to the queue. I realized that in practice it does not make sense for the queue to handle timeouts. The most import reason is that in order for a timeout to be successful, the process needs to be canceled. In reality this is in most cases not possible. If people want to handle timeouts they need to use a library that supports timeouts (like requestify) or they need to code it themselves. Promises can not be cancelled so in some cases you are screwed anyway. Another reason is that timeouts from the perspective of a queue are very hard to implement. Using `getTimeout` becomes tricky very fast. If the transaction happens to be a millisecond early the timeout will never happen.