added some extra unit tests to get to a 99% coverage of functionality
These were found by commenting out or replacing statements/expressions and checking to see if it affected existing tests.
Note that the 1% left is a memory optimization that allows the garbage collector to claim memory after shutdown. As far as I can tell there is no way to add a test for this effect.