cleanup of specs related code - still quite some work to be done there
It is interesting to see how this code has grown. Even after all the steps I took to reach 100% coverage I am still not sure if I have reached it. It is very hard to remain implementation independent in the tests. To safe the amount of code needed to test I removed some tests that exposed things covered in other tests. I probably need to go back and express them as separate tests again in order for it to be clear what behavior breaks when changing specific code. The ratio tests / production code will be insane (in my personal opinion) but I guess for a library that should be top-notch it is ok.