fix: handle non-object task types
The queue would crash if a task of the wrong type would come in. While it is a usage problem it
should not have undefined behavior. We can not leave it in the queue because it would continue to be
picked up when specs has `null` as `startState`. This commit contains one way to fix the problem. We
have to double check if this is the way we want to solve it.